Move around code and change wavefunction type signature
